Refactored resource packages to singular file
